+++ NEWS.md
@@ -3,6 +3,7 @@
  * Show diagnostic message, if the backend request fails in the places
   * Dashboard: in all three info boxes.
   * Tickets: Examine Ticket search field, events table and recipient info box.
+  * Stats: when getting Subqueries, searching for tickets/events, event details.
 
 
 ## 0.7.0 to 0.8.0